[~] Use half points for default grading schemes [~] Optimize imports

+++ src/packages/schoolexam/lib/exams/persistence/grading_table_lower_bound_data.dart
@@ -21,6 +21,6 @@ class GradingTableLowerBoundData {
 
   static GradingTableLowerBoundData fromMap(Map<String, dynamic> data) {
     return GradingTableLowerBoundData(
-        grade: data['grade'], points: data['points'], examId: data['examId']);
+        grade: data['grade'], points: data['points'].toDouble(), examId: data['examId']);
   }
 }
